Move the chess package under games/
Chess is about to stop being the only game in the project, so it moves from
game/ to games/chess/ to leave room for siblings.

It nests under games/ rather than sitting at the top level because a top-level
chess/ package would shadow python-chess, which ai/benchmark.py imports as
`import chess` to cross-check move generation.

Pure file move and import rewrite. No logic changes.
