+++ games/chess/constants.py
